+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes
 0.9 (unreleased)
 ----------------
 
+ - launchpad #1142994: url rewrites for vcs polling
+ - launchpad #1154673: treeStableTimer scheduler parameter now
+   configurable on a per-buildout basis
 
 0.8.1
 -----